+++ src/share/application_launcher.hpp
@@ -1,11 +1,17 @@
 #pragma once
 
 #include <cstdlib>
+#include <spdlog/fmt/fmt.h>
 #include <sstream>
 
 namespace krbn {
 class application_launcher final {
 public:
+  static void launch_app_icon_switcher(int number) {
+    auto command = fmt::format("'/Library/Application Support/org.pqrs/Karabiner-Elements/Karabiner-AppIconSwitcher.app/Contents/MacOS/Karabiner-AppIconSwitcher' {:03} &", number);
+    system(command.c_str());
+  }
+
   static void launch_event_viewer(void) {
     system("open '/Applications/Karabiner-EventViewer.app'");
   }
@@ -18,7 +24,7 @@ public:
     system("killall Karabiner-Menu");
   }
 
-  static void launch_preferences(void) {
+  static void launch_settings(void) {
     system("open '/Applications/Karabiner-Elements.app'");
   }
 
@@ -54,7 +60,7 @@ public:
   }
 
   static void launch_uninstaller(void) {
-    // Use nohup because uninstaller kill the Preferences Window.
+    // Use nohup because uninstaller kill the Settings Window.
     system("/usr/bin/nohup osascript '/Library/Application Support/org.pqrs/Karabiner-Elements/scripts/uninstaller.applescript' >/dev/null 2>&1 &");
   }
 };